moved testing folder and setup Travis
