+++ amz_review_scraper/amzscraper.py
@@ -9,16 +9,6 @@ from config import proxies
 from app import db
 import models
 
-#proxies = { 'http' : 'http://lum-customer-hl_9cbcc90f-zone-static:0rtgyzpvghf7@zproxy.lum-superproxy.io:22225',
-#           'https': 'https://lum-customer-hl_9cbcc90f-zone-static:0rtgyzpvghf7@zproxy.lum-superproxy.io:22225'
-#    }
-
-# For ignoring SSL certificate errors
-
-#ctx = ssl.create_default_context()
-#ctx.check_hostname = False
-#ctx.verify_mode = ssl.CERT_NONE
-
 header = {'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X x.y; rv:42.0) Gecko/20100101 Firefox/42.0'}
